feat: Request new scrape if scrape is seemingly frozen
If a scrape takes longer than 10 minutes, it has definitely frozen. When this happens... nothing happens. The user is never informed and the scrape is treated as still ongoing.
This changes that by checking if the scrape started a reasonable time ago, as well as letting the user know if the scrape has taken too long.