style: Move aspect ratio for cover image to wrapper
Images behave different when used with aspect-ratio. This is because their actual box content varies from their actual content (the image itself). The container might have the correct aspect ratio, but the image doesn't.

The aspect ratio is moved to the parent and the image is given object-fit to cover the parent.