refactor: Use picture tags for theme images
Previously images would only load correctly if the theme cookie was set. This was never the case on first load. By using a picture tag we get the use `source` tags with a `media` attribute for each theme image.

As a bonus point, this is a million times easier to read.